requirements.txt
Flask==2.3.2
gunicorn==21.2.0
google-cloud-datastore==2.18.0
matplotlib==3.8.4
